+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is inspired from [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and the versioning aim to respect [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- temp dir for datapackages
+- query parameters in upload_datapackage
+
 ## [0.9.0] - 2022-06-15
 ### Added
 - query page with dynamic source parameters